TODO: update task

+++ scripts/common.js
@@ -11,28 +11,35 @@ $('#tasks-container').on('click', '.done-checkbox', function () {
         isDone: $(this).prop('checked')
     };
     
-    taskService.updateTask(task);
+    //taskService.updateTask(task);
 });
 
 $('#tasks-container').on('click', '.remove', function () {
-    $(this).parents('.task-row').remove();
-    var taskId = $(this).parents('.task-row').data('taskid');    
-    taskService.deleteTask(taskId);
+    // get id of a task that stored in data attribute data-taskid
+    var taskId = $(this).parents('.task-row').data('taskid');
+
+    // remove .task-row from DOM
+    $(this).parents('.task-row').remove();  
+    
+    // if id of a task is not undefined call the service deleteTask method
+    if (taskId != undefined) {
+        taskService.deleteTask(taskId);
+    }   
 });
 
 $('#save-btn').click(function () {
-    var container = $('#tasks-container');
-    var title = $('#task').val();    
+    // get input value
+    var title = $('#task').val();
+
+    // if input value is not empty - create task object and call service addTask method
     if (title.trim() != '') {
         var task = {
             title: title
-        };
-
-        //var element = buildTask(task);
-        //container.append(element);      
+        };             
 
         taskService.addTask(task);
 
+        // clear an input value
         $('#task').val('');
     } else {
         showErrorMessage('Please enter a task title');
@@ -44,11 +51,21 @@ $('#task').keydown(function () {
     $('.error-message').remove();
 })
 
+
+/* function that creates new task row element and appends it to the tasks container
+ * input parameter: object task
+ */
 function buildTask(task) {
-    var el = '<div class="checkbox task-row'+ (task.isDone ? ' completed' : '') +'"data-taskid='+task.id+'><label><input type="checkbox" value="" class="done-checkbox">' + task.title + '</label>';
+    // tasks container
+    var container = $('#tasks-container');
+
+    // element which will be appended to the tasks container
+    var el = '<div class="checkbox task-row'+ (task.isDone ? ' completed' : '') +'" data-taskid='+ task.id +'>';
+    el += '<label><input type="checkbox" value="" class="done-checkbox">' + task.title + '</label>';
     el += '<i class="fa fa-times remove pull-right" aria-hidden="true"></i></div>';
 
-    return el;
+    // append just created element to the tasks container
+    container.append(el);
 }
 
 function showErrorMessage(message) {
